+++ src/brushes/BaseBrush.ts
@@ -2,7 +2,7 @@ import { Color } from '../color/Color';
 import type { Point } from '../Point';
 import { TEvent } from '../EventTypeDefs';
 import type { Shadow } from '../Shadow';
-import type { Canvas } from '../canvas/CanvasEvents';
+import type { Canvas } from '../canvas/Canvas';
 
 type TBrushEventData = TEvent & { pointer: Point };
