+++ src/brushes/CircleBrush.ts
@@ -4,7 +4,7 @@ import { Shadow } from '../Shadow';
 import { Circle } from '../shapes/Circle';
 import { Group } from '../shapes/Group';
 import { getRandomInt } from '../util/internals';
-import type { Canvas } from '../canvas/CanvasEvents';
+import type { Canvas } from '../canvas/Canvas';
 import { BaseBrush } from './BaseBrush';
 
 export type CircleBrushPoint = {
