+++ src/brushes/SprayBrush.ts
@@ -3,7 +3,7 @@ import { Group } from '../shapes/Group';
 import { Shadow } from '../Shadow';
 import { Rect } from '../shapes/Rect';
 import { getRandomInt } from '../util/internals';
-import type { Canvas } from '../canvas/CanvasEvents';
+import type { Canvas } from '../canvas/Canvas';
 import { BaseBrush } from './BaseBrush';
 
 export type SprayBrushPoint = {
